+++ gui/wxpython/mapdisp/frame.py
@@ -142,7 +142,38 @@ class MapFrame(SingleMapFrame):
         #
         self.statusbarManager = None
         if statusbar:
-            self.statusbar = self.CreateStatusbar()
+            # items for choice
+            statusbarItems = [
+                sb.SbCoordinates,
+                sb.SbRegionExtent,
+                sb.SbCompRegionExtent,
+                sb.SbShowRegion,
+                sb.SbAlignExtent,
+                sb.SbResolution,
+                sb.SbDisplayGeometry,
+                sb.SbMapScale,
+                sb.SbGoTo,
+                sb.SbProjection,
+            ]
+            self.statusbarItemsHiddenInNviz = (
+                sb.SbAlignExtent,
+                sb.SbDisplayGeometry,
+                sb.SbShowRegion,
+                sb.SbResolution,
+                sb.SbMapScale,
+            )
+            self.statusbar = self.CreateStatusbar(statusbarItems)
+
+            self.Map.GetRenderMgr().updateProgress.connect(
+                self.statusbarManager.SetProgress
+            )
+
+        self.Map.GetRenderMgr().renderingFailed.connect(
+            lambda cmd, error: self._giface.WriteError(
+                _("Failed to run command '%(command)s'. Details:\n%(error)s")
+                % dict(command=" ".join(cmd), error=error)
+            )
+        )
 
         # init decoration objects
         self.decorations = {}
@@ -208,20 +239,9 @@ class MapFrame(SingleMapFrame):
             .Layer(0),
         )
 
-        self._mgr.AddPane(
-            self.statusbar,
-            wx.aui.AuiPaneInfo()
-            .Bottom()
-            .MinSize(30, 30)
-            .Fixed()
-            .Name("statusbar")
-            .CloseButton(False)
-            .DestroyOnClose(True)
-            .ToolbarPane()
-            .Dockable(False)
-            .PaneBorder(False)
-            .Gripper(False),
-        )
+        # statusbar
+        self.AddStatusbarPane()
+
         self._mgr.Update()
 
         #
@@ -279,75 +299,6 @@ class MapFrame(SingleMapFrame):
             label=show_hide_statusbar_label,
             action=on_show_hide_statusbar,
         )
-
-    def CreateStatusbar(self):
-        if self.statusbarManager:
-            return
-
-        # items for choice
-        self.statusbarItems = [
-            sb.SbCoordinates,
-            sb.SbRegionExtent,
-            sb.SbCompRegionExtent,
-            sb.SbShowRegion,
-            sb.SbAlignExtent,
-            sb.SbResolution,
-            sb.SbDisplayGeometry,
-            sb.SbMapScale,
-            sb.SbGoTo,
-            sb.SbProjection,
-        ]
-
-        self.statusbarItemsHiddenInNviz = (
-            sb.SbAlignExtent,
-            sb.SbDisplayGeometry,
-            sb.SbShowRegion,
-            sb.SbResolution,
-            sb.SbMapScale,
-        )
-
-        statusbar = wx.StatusBar(self, id=wx.ID_ANY)
-        statusbar.SetMinHeight(24)
-        statusbar.SetFieldsCount(4)
-        statusbar.SetStatusWidths([-5, -2, -1, -1])
-        self.statusbarManager = sb.SbManager(mapframe=self, statusbar=statusbar)
-
-        # fill statusbar manager
-        self.statusbarManager.AddStatusbarItemsByClass(
-            self.statusbarItems, mapframe=self, statusbar=statusbar
-        )
-        self.statusbarManager.AddStatusbarItem(
-            sb.SbMask(self, statusbar=statusbar, position=2)
-        )
-        sbRender = sb.SbRender(self, statusbar=statusbar, position=3)
-        self.statusbarManager.AddStatusbarItem(sbRender)
-
-        self.statusbarManager.Update()
-
-        #
-        self.Map.GetRenderMgr().updateProgress.connect(
-            self.statusbarManager.SetProgress
-        )
-        self.Map.GetRenderMgr().renderingFailed.connect(
-            lambda cmd, error: self._giface.WriteError(
-                _("Failed to run command '%(command)s'. Details:\n%(error)s")
-                % dict(command=" ".join(cmd), error=error)
-            )
-        )
-        return statusbar
-
-    def ShowStatusbar(self, show):
-        """Show/hide statusbar and associated pane"""
-        self._mgr.GetPane("statusbar").Show(show)
-        self._mgr.Update()
-
-    def IsStatusbarShown(self):
-        """Check if statusbar is shown"""
-        return self._mgr.GetPane("statusbar").IsShown()
-
-    def SetStatusText(self, *args):
-        """Overide wx.StatusBar method"""
-        self.statusbar.SetStatusText(*args)
 
     def GetMapWindow(self):
         return self.MapWindow
@@ -699,13 +650,7 @@ class MapFrame(SingleMapFrame):
         :param name toolbar to remove
         :param destroy True to destroy otherwise toolbar is only hidden
         """
-        self._mgr.DetachPane(self.toolbars[name])
-        if destroy:
-            self._toolSwitcher.RemoveToolbarFromGroup("mouseUse", self.toolbars[name])
-            self.toolbars[name].Destroy()
-            self.toolbars.pop(name)
-        else:
-            self.toolbars[name].Hide()
+        super().RemoveToolbar(name, destroy)
 
         if name == "vdigit":
             self._mgr.GetPane("vdigit").Hide()
@@ -715,23 +660,6 @@ class MapFrame(SingleMapFrame):
         self.toolbars["map"].Enable2D(True)
 
         self._mgr.Update()
-
-    def ShowAllToolbars(self, show=True):
-        if not show:  # hide
-            action = self.RemoveToolbar
-        else:
-            action = self.AddToolbar
-        for toolbar in self.GetToolbarNames():
-            action(toolbar)
-
-    def AreAllToolbarsShown(self):
-        return self.GetMapToolbar().IsShown()
-
-    def IsPaneShown(self, name):
-        """Check if pane (toolbar, mapWindow ...) of given name is currently shown"""
-        if self._mgr.GetPane(name).IsOk():
-            return self._mgr.GetPane(name).IsShown()
-        return False
 
     def RemoveQueryLayer(self):
         """Removes temporary map layers (queries)"""
@@ -1649,10 +1577,6 @@ class MapFrame(SingleMapFrame):
     def GetMapToolbar(self):
         """Returns toolbar with zooming tools"""
         return self.toolbars["map"] if "map" in self.toolbars else None
-
-    def GetToolbarNames(self):
-        """Return toolbar names"""
-        return self.toolbars.keys()
 
     def GetDialog(self, name):
         """Get selected dialog if exist"""
